refactor(Name): Player 이름 멤버 변수 클래스로 분리
